fix(模型训练): 修复训练过程中的梯度计算和模型状态问题
- 将diffusion计算移出对比学习循环，避免重复计算
- 为teacher_embed添加模型状态管理，确保评估模式
- 调整delta_times的预处理方式，增加数值稳定性
- 在归一化操作中增加最小epsilon值防止除零错误

+++ methods/contrastive_transformer_teacher_diffusion/create_embeddings.py
@@ -118,10 +118,10 @@ def train_model(
                 s1 = model.student_forward(v1)
                 s2 = model.student_forward(v2)
                 ctr_loss = info_nce_loss(s1["embed"], s2["embed"], temperature=temperature)
-                base_hidden = model.student_encoder(batch)
-                teacher_target = model.teacher_embed(batch)
-                diff_loss = model.diffusion_loss(base_hidden, teacher_target)
-                loss = contrastive_weight * ctr_loss + diffusion_weight * diff_loss
+            base_hidden = model.student_encoder(batch)
+            teacher_target = model.teacher_embed(batch).to(base_hidden.dtype)
+            diff_loss = model.diffusion_loss(base_hidden, teacher_target)
+            loss = contrastive_weight * ctr_loss + diffusion_weight * diff_loss
             scaler.scale(loss).backward()
             scaler.step(optimizer)
             scaler.update()
